+++ src/migrations/m190614_103420_AddMissingMetaColumnsToProAndPaymentTables.php
@@ -3,6 +3,7 @@
 namespace Solspace\Freeform\migrations;
 
 use craft\db\Migration;
+use yii\db\Exception;
 
 /**
  * m190614_103420_AddMissingMetaColumnsToPaymentTables migration.
@@ -14,39 +15,21 @@ class m190614_103420_AddMissingMetaColumnsToProAndPaymentTables extends Migratio
      */
     public function safeUp()
     {
-        try {
-            $this->addColumn('{{%freeform_export_profiles}}', 'dateCreated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_export_profiles}}', 'dateUpdated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_export_profiles}}', 'uid', $this->uid());
-        } catch (\Exception $e) {
-        }
+        $tables = [
+            'freeform_export_profiles',
+            'freeform_export_settings',
+            'freeform_payments_subscription_plans',
+            'freeform_payments_payments',
+            'freeform_payments_subscriptions',
+        ];
 
-        try {
-            $this->addColumn('{{%freeform_export_settings}}', 'dateCreated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_export_settings}}', 'dateUpdated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_export_settings}}', 'uid', $this->uid());
-        } catch (\Exception $e) {
-        }
-
-        try {
-            $this->addColumn('{{%freeform_payments_subscription_plans}}', 'dateCreated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_payments_subscription_plans}}', 'dateUpdated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_payments_subscription_plans}}', 'uid', $this->uid());
-        } catch (\Exception $e) {
-        }
-
-        try {
-            $this->addColumn('{{%freeform_payments_payments}}', 'dateCreated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_payments_payments}}', 'dateUpdated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_payments_payments}}', 'uid', $this->uid());
-        } catch (\Exception $e) {
-        }
-
-        try {
-            $this->addColumn('{{%freeform_payments_subscriptions}}', 'dateCreated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_payments_subscriptions}}', 'dateUpdated', $this->dateTime()->notNull());
-            $this->addColumn('{{%freeform_payments_subscriptions}}', 'uid', $this->uid());
-        } catch (\Exception $e) {
+        foreach ($tables as $table) {
+            try {
+                $this->addColumn("{{%$table}}", 'dateCreated', $this->dateTime());
+                $this->addColumn("{{%$table}}", 'dateUpdated', $this->dateTime());
+                $this->addColumn("{{%$table}}", 'uid', $this->uid());
+            } catch (\Exception $e) {
+            }
         }
 
         return true;